requirements.txt
cbor2==5.6.5
future==1.0.0
iso8601==2.1.0
luma.core==2.4.2
luma.lcd==2.11.0
luma.oled==3.14.0
pillow==11.1.0
pyftdi==0.56.0
pypng==0.20220715.0
pyserial==3.5
pyusb==1.3.1
PyYAML==6.0.2
RPi.GPIO==0.7.1
serial==0.0.97
smbus==1.1.post2
smbus2==0.5.0
spidev==3.6
